+++ test/testData.js
@@ -1,50 +1,12 @@
 // @flow
 
-interface User {
-  id: string,
-  email: string,
-  passwordHash: string,
-  alias: string,
-}
-
-interface ApiToken {
-  token: string,
-  accountName: string,
-  accountId: string,
-  permissions: string,
-  world: number,
-  guilds: string,
-  User: User,
-  UserId: string,
-}
-
-interface Guild {
-  id: string,
-  tag: string,
-  name: string,
-  favor: number,
-  resonance: number,
-  aetherium: number,
-  influence: number,
-  level: number,
-  motd: string,
-  claimed: boolean,
-  apiToken?: ?string,
-}
-
-interface Character {
-  name: string,
-  race: string,
-  gender: 'Female' | 'Male',
-  profession: string,
-  level: number,
-  created: string,
-  age: number,
-  deaths: number,
-  guild: string,
-  Gw2ApiTokenToken: string,
-  Gw2ApiToken: ApiToken,
-}
+import type {
+  PvpStanding,
+  User,
+  Guild,
+  ApiToken,
+  Character,
+} from 'flowTypes';
 
 const defaultUser: User = {
   id: '938C506D-F838-F447-8B43-4EBF34706E0445B2B503',
@@ -93,21 +55,48 @@ export const apiToken = (apiTkn: ?ApiToken): ApiToken => ({
   ...apiTkn,
 });
 
-const defaultCharacter: Character = {
-  name: 'character',
-  race: 'race',
-  gender: 'Male',
-  profession: 'Elementalist',
-  level: 20,
-  created: '01/02/1990',
-  age: 30,
-  deaths: 10,
-  guild: guild().name,
-  Gw2ApiTokenToken: apiToken().token,
-  Gw2ApiToken: apiToken(),
+export const character = ({
+  name = 'character',
+  race = 'race',
+  gender = 'Male',
+  profession = 'Elementalist',
+  level = 20,
+  created = '01/02/1990',
+  age = 30,
+  deaths = 10,
+  guild: gld = guild().name,
+  Gw2ApiTokenToken = apiToken().token,
+  Gw2ApiToken = apiToken(),
+}: Character = {}): Character => ({
+  name,
+  race,
+  gender,
+  profession,
+  level,
+  created,
+  age,
+  deaths,
+  guild: gld,
+  Gw2ApiTokenToken,
+  Gw2ApiToken,
+});
+
+const fakeStanding = {
+  totalPoints: 26,
+  division: 0,
+  tier: 1,
+  points: 6,
+  repeats: 0,
+  rating: 1159,
 };
 
-export const character = (char: ?Character): Character => ({
-  ...defaultCharacter,
-  ...char,
+export const standing = ({
+  current = fakeStanding,
+  best = fakeStanding,
+  seasonId = 'A54849B7-7DBD-4958-91EF-72E18CD659BA',
+
+}: PvpStanding): PvpStanding => ({
+  current,
+  best,
+  seasonId,
 });