+++ CHANGELOG.md
@@ -8,6 +8,14 @@
 - Changes
 ```
 
+## [2.0.0] - unreleased
+### Added
+- PvpStandings table
+- Pvp standings data fetcher
+
+### Changed
+- Gw2 lib to normalise pvp standings data
+
 ## [1.0.0] - 2017/01/02
 ### Added
 - This CHANGELOG file to allow more insight to the changes made throughout the development of api.gw2armory.com